db/requirements.txt
psycopg2-binary>=2.9.5
sqlalchemy>=2.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
cloud-sql-python-connector>=1.2.0
pg8000>=1.30.0
pandas>=1.3.0
alembic>=1.11.0
pytest>=7.3.1
